refactor(timeline): 将日程体系重构为 Timeline 关键帧体系
删除旧的 Schedule/ScheduleDuties/DutyTime 体系，新建 world/attribute 与
world/timeline 模块，复刻 MC 1.21.11 的 Timeline 采样语义。

- world/attribute：LerpFunction、AttributeRange、AttributeType、
  AttributeModifier、EnvironmentAttribute(Layer)，预定义 ACTIVITY 类型与
  VILLAGER_ACTIVITY/BABY_VILLAGER_ACTIVITY 属性
- world/timeline：Keyframe、EasingType、KeyframeTrack(Sampler)、
  AttributeTrack(Sampler)、Timeline，预定义 VILLAGER_SCHEDULE
- Brain：m_schedulePtr 改为 m_timelinePtr 与活动轨道采样器，
  _updateActivity 按 dayTime 采样当前活动
- VillagerEntity：按成年/幼年注册对应活动属性
- RegistryBootstrap：Schedule::initialize() 改为 Timelines::bootstrap()
- 删除 Schedule.hpp/Schedule.cpp/DutyTime.hpp（Activity 保留）
- 测试：新增 TimelineTest，原 BrainScheduleTests 拆为 BrainMemoryTests

关键语义：活动为离散值采用阶跃插值；周期回绕段使 dayTime 早于首个关键帧时
回落到上一周期末值；预定义时间线列举顺序与网络层 minecraft:timeline
注册表条目一致。

+++ src/common/world/attribute/EnvironmentAttributeLayer.hpp
@@ -23,6 +23,8 @@
 
 #pragma once
 
+#include "common/core/Types.hpp"
+
 namespace mc {
 namespace world {
 namespace attribute {
@@ -57,7 +59,7 @@ public:
     class TimeBased {
     public:
         virtual ~TimeBased() = default;
-        virtual Value applyTimeBased(Value value, int tick) = 0;
+        virtual Value applyTimeBased(Value value, i64 tick) = 0;
     };
 
     /**
